"check_style_score -> 5 of 5 passed. style score: 0/1 "

+++ CSPP2-Practice/module 2/Assignment-3/Solution.java
@@ -12,13 +12,13 @@ public class Solution {
 		Scanner s=new Scanner(System.in);      
         int base = s.nextInt();
         int exponent = s.nextInt();
-        int result=power(base,exponent);
+        long result=power(base,exponent);
         System.out.println(result);
 	}
 	/*
 	Need to write the power function and print the output.
 	*/
-	public static int power(int base, int exponent) {
+	public static long power(int base, int exponent) {
 		if (exponent == 0)	return 1;
 		else if (exponent == 1) return base;
 		else return base * power(base, exponent-1);
